+++ .storybook/preview.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import type { Preview } from "@storybook/react";
 import UIThemeProvider from "../src/v1/theme/UIThemeProvider";
 import { UIThemeSchema } from "../src/v1/theme/colors/theme-colors";
+import "../src/assets/fonts/fontawesome/css/all.min.css";
 
 const preview: Preview = {
   decorators: [